requirements.txt
torch
torchvision
opencv-python
tensorflow <=2.15.0
google-generativeai
PyYAML==6.0.1
pycocotools
pybboxes
pillow<=9.5.0
timm
dvc
Flask
Flask-cors
gdown
mlflow>=2.5.0
dagshub
setuptools~=69.2.0
-e .